feat: make `expiration_date` explicit and required
