+++ agent_issuance/src/credential/command.rs
@@ -1,8 +1,7 @@
-use agent_shared::config::CredentialExpiry;
 use oid4vci::credential_issuer::credential_configurations_supported::CredentialConfigurationsSupportedObject;
 use serde::Deserialize;
 
-use super::entity::Data;
+use super::{aggregate::CredentialExpiry, entity::Data};
 
 #[derive(Debug, Deserialize)]
 #[serde(untagged)]
@@ -11,7 +10,7 @@ pub enum CredentialCommand {
         credential_id: String,
         data: Data,
         credential_configuration: Box<CredentialConfigurationsSupportedObject>,
-        expires: Option<CredentialExpiry>,
+        expires: CredentialExpiry,
     },
     CreateSignedCredential {
         credential_id: String,
