+++ agent_shared/src/config.rs
@@ -27,48 +27,10 @@ pub struct ApplicationConfiguration {
     pub secret_manager: SecretManagerConfig,
     pub did_document_cache: Option<InMemoryCacheConfig>,
     pub credential_configurations: Vec<CredentialConfiguration>,
-    #[serde(default = "default_credential_expiry")]
-    pub credential_expiry: CredentialExpiry,
     pub signing_algorithms_supported: HashMap<jsonwebtoken::Algorithm, ToggleOptions>,
     pub display: Vec<Display>,
     pub event_publishers: Option<EventPublishers>,
     pub vp_formats: HashMap<ClaimFormatDesignation, ToggleOptions>,
-}
-
-fn default_credential_expiry() -> CredentialExpiry {
-    CredentialExpiry::Never
-}
-
-#[derive(Debug, Serialize, Deserialize, Clone, PartialEq)]
-#[serde(untagged)]
-pub enum CredentialExpiry {
-    Fixed(chrono::DateTime<chrono::Utc>),
-    Relative(iso8601_duration::Duration),
-    #[serde(with = "never_as_str")]
-    Never,
-}
-
-mod never_as_str {
-    use serde::{Deserialize, Deserializer, Serializer};
-
-    pub fn serialize<S>(serializer: S) -> Result<S::Ok, S::Error>
-    where
-        S: Serializer,
-    {
-        serializer.serialize_str("never")
-    }
-
-    pub fn deserialize<'de, D>(deserializer: D) -> Result<(), D::Error>
-    where
-        D: Deserializer<'de>,
-    {
-        let s = String::deserialize(deserializer)?;
-        if s == "never" {
-            Ok(())
-        } else {
-            Err(serde::de::Error::custom("expected 'never'"))
-        }
-    }
 }
 
 #[derive(Debug, Deserialize, Clone, Default)]
@@ -446,14 +408,5 @@ mod tests {
         for variant in SupportedDidMethod::VARIANTS {
             let _subject_syntax_type: SubjectSyntaxType = variant.clone().into();
         }
-    }
-
-    #[test]
-    fn custom_serializer_for_credential_expiry() {
-        let deserialized: CredentialExpiry = serde_json::from_value(serde_json::json!("never")).unwrap();
-        assert_eq!(deserialized, CredentialExpiry::Never);
-
-        let serialized = serde_json::to_value(&CredentialExpiry::Never).unwrap();
-        assert_eq!(serialized, serde_json::json!("never"));
     }
 }
